hotfix: - 컴파일 가능하도록 변경, 디렉토리 변경
