+++ Look-out-the-window/Look-out-the-window/Data/Model/DetailModel.swift
@@ -1,12 +1,11 @@
 //
-//  MainSectionDataModel.swift
+//  DetailModel.swift
 //  Look-out-the-window
 //
-//  Created by GO on 5/21/25.
+//  Created by 서동환 on 5/27/25.
 //
 
-import UIKit
-import RxDataSources
+import Foundation
 
 enum DetailType {
     case sunriseSunset
@@ -61,70 +60,7 @@ enum DetailType {
     }
 }
 
-/// Header
-enum SectionHeaderInfo: Int, CaseIterable {
-    case hourly
-    case daily
-
-    var icon: String {
-        switch self {
-        case .hourly: return "clock"
-        case .daily: return "calendar"
-        }
-    }
-
-    var title: String {
-        switch self {
-        case .hourly: return "시간별 예보"
-        case .daily: return "일별 예보"
-        }
-    }
-}
-
-
-// MARK: - 각 섹션 DataModel
-struct HourlyModel: Hashable {
-    let hour: Int
-    let temperature: String
-    // weatherState
-    let weatherInfo: String // Asset네이밍 변환받아 전달받을 예정
-}
-
-struct DailyModel: Hashable {
-    // 요일, 하늘상태(이미지 -> String), 최저 - 최고 온도
-    let unixTime: Int
-    let day: String
-    let high: String
-    let low: String
-    let weatherInfo: String // Asset네이밍 변환받아 전달받을 예정
-    let maxTemp: Int
-    let minTemp: Int
-//    let temperature: String // 현재 온도
-}
-
 struct DetailModel {
     let title: DetailType
     let value: String
 }
-
-enum MainSectionItem {
-    case hourly(HourlyModel)
-    case daily(DailyModel)
-    case detail(DetailModel)
-}
-
-struct MainSection {
-    var items: [MainSectionItem]
-}
-
-extension MainSection: SectionModelType {
-    typealias Item = MainSectionItem
-
-    init(original: MainSection, items: [MainSectionItem]) {
-        self = original
-        self.items = items
-    }
-}
-
-
-
