+++ Look-out-the-window/Look-out-the-window/Presentation/Common/View/BackgroundTopInfoView.swift
