--- Look-out-the-window/Look-out-the-window/Presentation/Main/View/RiveNames.swift
@@ -1,18 +1,0 @@
-//
-//  RiveNames.swift
-//  Look-out-the-window
-//
-//  Created by 정근호 on 5/21/25.
-//
-
-import Foundation
-
-enum Rive {
-    static let sunny = "Sunny"
-    static let partlyCloudy = "PartlyCloudy"
-    static let cloudy = "Cloudy"
-    static let rainy = "Rainy"
-    static let snow = "Snow"
-    static let thunder = "Thunderbolt"
-    static let fog = "Fog"
-}
